receipt added

+++ EczaneDepoDB/Modal/DataAccess/Receipt.cs
@@ -7,18 +7,17 @@ namespace EczaneDepoDB.Modal.DataAccess
     {
         public int Id { get; set; }
         public int PrescriptionId { get; set; }
-        public int Price { get; set; }
+        public decimal Price { get; set; }
         public int PhoneNumber { get; set; }
 
-
-        public Receipt(int id, int prescriptionId, int price, int phoneNumber)
+        public Receipt(int id, int prescriptionId, decimal price, int phoneNumber)
         {
             Id = id;
             PrescriptionId = prescriptionId;
             Price = price;
             PhoneNumber = phoneNumber;
         }
-        public Receipt(int prescriptionId, int price, int phoneNumber)
+        public Receipt(int prescriptionId, decimal price, int phoneNumber)
         {
             PrescriptionId = prescriptionId;
             Price = price;
